server/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
nbformat
nbconvert
pypdf
python-magic